refactor: encode/decode
* rename charArray to alphabet

* move alphabets out of static members into a separate module
because there is no reason to bound alphabets to the BitArray class

* rename encodeWithCharset/decodeWithCharset to encode/decode
because the method name should not describe the arguments.
Instead the argument names should be self-descriptive.
Besides, we may want to make charset/alphabet optional in the future
(with a default value).

* reorder the arguments in decode
This is more future-proof, should we want to have a default alphabet.

* simplify error message.