fix: file extension in import
No file extension in import statements work ok with typescript.
However, it will failt when this library is used as a dependency,
because the dependent will expect a .js extension to find alphabet.js.

This change goes for a quick fix, by removing the import altogether.
I should come with a more robust solution in the future.